platforms/python/django-react-app/requirements.txt
# Note: Sentry SDK will be installed separately

Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
datadog==0.44.0
whitenoise==6.2.0  # serving static files

ipdb  # debugging
